Final Class Design
