+++ src/main/java/com/modisa/classdesign/models/classdesign/encapsulation/appcanvas/Canvas.java
@@ -1,0 +1,10 @@
+package com.modisa.classdesign.models.classdesign.encapsulation.appcanvas;
+
+import com.modisa.classdesign.models.classdesign.encapsulation.graphicshape.Circle;
+
+public class Canvas {
+    void getArea() {
+        Circle circle = new Circle();
+        circle.area(); //call to public method area(), outside package
+    }
+}
