+++ src/main/java/com/modisa/classdesign/models/classdesign/encapsulation/graphicshape/Shape.java
@@ -1,0 +1,5 @@
+package com.modisa.classdesign.models.classdesign.encapsulation.graphicshape;
+
+public class Shape {
+    protected int color;
+}
