+++ src/main/java/com/modisa/classdesign/models/classdesign/inheritance/TestNumber.java
@@ -1,0 +1,13 @@
+package com.modisa.classdesign.models.classdesign.inheritance;
+
+public class TestNumber {
+    // Take array of numbers and sum them up
+    public static double sum(Number[] numbs) {
+        double sum = 0d;
+
+        for (Number num: numbs) {
+            sum += num.doubleValue();
+        }
+        return sum;
+    }
+}
