+++ src/main/java/com/modisa/classdesign/models/classdesign/polymorphism/dynamicpoly/Circle.java
@@ -1,0 +1,11 @@
+package com.modisa.classdesign.models.classdesign.polymorphism.dynamicpoly;
+
+public class Circle extends Shape {
+    private int radius;
+    public Circle(int radius) {
+        this.radius = radius;
+    }
+    public double area() {
+        return Math.PI * radius * radius;
+    }
+}
