+++ src/main/java/com/modisa/classdesign/models/classdesign/polymorphism/dynamicpoly/Shape.java
@@ -1,0 +1,8 @@
+package com.modisa.classdesign.models.classdesign.polymorphism.dynamicpoly;
+
+public class Shape {
+    public double area() {
+        return 0;
+    }
+
+}
