+++ src/main/java/com/modisa/classdesign/models/classdesign/polymorphism/staticpoly/Calculate.java
@@ -1,0 +1,19 @@
+package com.modisa.classdesign.models.classdesign.polymorphism.staticpoly;
+
+public class Calculate {
+    public double add(int a, int b) {
+        return a + b;
+    }
+
+    public double add(int a, int b, int c) {
+        return a + b + c;
+    }
+
+    public double add(double a, int b) {
+        return a + b;
+    }
+
+    public double add(int a, double b) {
+        return a + b;
+    }
+}
