+++ src/main/java/com/modisa/classdesign/models/classdesign/statickeyword/Counter.java
@@ -1,0 +1,13 @@
+package com.modisa.classdesign.models.classdesign.statickeyword;
+
+public class Counter {
+    private static int count;
+
+    public Counter() {
+        count++;
+    }
+
+    public void printCount() {
+        System.out.println("Number of instances created so far is: " + count);
+    }
+}
