--- src/main/java/com/modisa/classdesign/models/encapsulation/appcanvas/Canvas.java
@@ -1,10 +1,0 @@
-package com.modisa.classdesign.models.encapsulation.appcanvas;
-
-import com.modisa.classdesign.models.encapsulation.graphicshape.Circle;
-
-public class Canvas {
-    void getArea() {
-        Circle circle = new Circle();
-        circle.area(); //call to public method area(), outside package
-    }
-}
