--- src/main/java/com/modisa/classdesign/models/encapsulation/graphicshape/Shape.java
@@ -1,5 +1,0 @@
-package com.modisa.classdesign.models.encapsulation.graphicshape;
-
-public class Shape {
-    protected int color;
-}
