--- src/main/java/com/modisa/classdesign/models/inheritance/TestNumber.java
@@ -1,13 +1,0 @@
-package com.modisa.classdesign.models.inheritance;
-
-public class TestNumber {
-    // Take array of numbers and sum them up
-    public static double sum(Number[] numbs) {
-        double sum = 0d;
-
-        for (Number num: numbs) {
-            sum += num.doubleValue();
-        }
-        return sum;
-    }
-}
