--- src/main/java/com/modisa/classdesign/models/overiding/object/clas/Point.java
@@ -1,18 +1,0 @@
-package com.modisa.classdesign.models.overiding.object.clas;
-
-public class Point {
-    private int xPos, yPos;
-
-    public Point(int xPos, int yPos) {
-        this.xPos = xPos;
-        this.yPos = yPos;
-    }
-
-    @Override
-    public String toString() {
-        return "Point{" +
-                "xPos=" + xPos +
-                ", yPos=" + yPos +
-                '}';
-    }
-}
