--- src/main/java/com/modisa/classdesign/models/polymorphism/dynamicpoly/Shape.java
@@ -1,8 +1,0 @@
-package com.modisa.classdesign.models.polymorphism.dynamicpoly;
-
-public class Shape {
-    public double area() {
-        return 0;
-    }
-
-}
