--- src/main/java/com/modisa/classdesign/models/polymorphism/dynamicpoly/Square.java
@@ -1,12 +1,0 @@
-package com.modisa.classdesign.models.polymorphism.dynamicpoly;
-
-public class Square extends Shape {
-    private int side;
-    public Square(int side) {
-        this.side = side;
-    }
-
-    public double area() {
-        return (side * side);
-    }
-}
